update dc_code and dc_law permalink modules
* rename the `dccode` linker to `dc_council`
* fix its dc_law linker function -- it was incorrectly applied to dc_code citations, and the URLs have changed to https://beta.code.dccouncil.us
* add a dc_code linker function to https://beta.code.dccouncil.us
* drop the dccodeorg and dcdecoded linkers since the sites are no longer maintained
